+++ packages/observable-client/src/chainHead/streams/pinned-blocks.ts
@@ -41,8 +41,14 @@ export type PinnedBlocks = {
 }
 
 const deleteBlock = (blocks: PinnedBlocks["blocks"], blockHash: string) => {
-  blocks.get(blocks.get(blockHash)!.parent)?.children.delete(blockHash)
-  blocks.delete(blockHash)
+  const block = blocks.get(blockHash)
+  // the following `if` statement is in -in theory- redundant, as the spec clearly states that:
+  // > All items in `finalizedBlockHashes` and `prunedBlockHashes` are guaranteed to have been reported through earlier newBlock events.
+  // However, some PolkadotSDK nodes send in their list of `prunnedBlocks` blocks that have not been previously reported.
+  if (block) {
+    blocks.get(block.parent)?.children.delete(blockHash)
+    blocks.delete(blockHash)
+  }
 }
 
 const getBlocksToUnpin = (blocks: PinnedBlocks, pruned: string[]) => {
@@ -63,7 +69,7 @@ const getBlocksToUnpin = (blocks: PinnedBlocks, pruned: string[]) => {
   const deletedBlocks = [...pruned]
   for (let i = trail.length - 1; i >= 0; i--) {
     current = blocks.blocks.get(trail[i])!
-    if (!current.unpinned) return result
+    if (!current.unpinned) break
     deletedBlocks.push(current.hash)
   }